pyproject.toml
[tool.poetry]
name = "genai-platform-services"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

# TO DO : Need to delete unused dependency.
[tool.poetry.dependencies]
python = "3.11.9"
litellm = "1.63.0"
aiohappyeyeballs = "2.5.0"
aiohttp = "3.11.13"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.8.0"
attrs = "25.1.0"
beautifulsoup4 = "4.13.3"
cachetools = "5.5.2"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
distro = "1.9.0"
fastapi = "0.115.11"
filelock = "3.17.0"
frozenlist = "1.5.0"
fsspec = {extras = ["aws","gcp"], version = "^2025.3.2"}
google = "3.0.0"
google-auth = "2.38.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.1"
h11 = "0.14.0"
httpcore = "1.0.7"
httplib2 = "0.22.0"
httpx = {extras = ["http2"], version = "^0.28.1"}
idna = "3.10"
importlib-metadata = "8.6.1"
jinja2 = "3.1.6"
jiter = "0.8.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
markupsafe = "3.0.2"
multidict = "6.1.0"
oauthlib = "3.2.2"
openai = "1.65.4"
packaging = "24.2"
propcache = "0.3.0"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pydantic = "2.10.6"
pydantic-core = "2.27.2"
pyparsing = "3.2.1"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
referencing = "0.36.2"
regex = "2024.11.6"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rpds-py = "0.23.1"
rsa = "4.9"
sniffio = "1.3.1"
soupsieve = "2.6"
starlette = "0.46.0"
tiktoken = "0.9.0"
tokenizers = "0.21.0"
tqdm = "4.67.1"
typing-extensions = "4.12.2"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
yarl = "1.18.3"
zipp = "3.21.0"
google-cloud-storage = "^3.1.0"
pymupdf4llm = "^0.0.21"
gcsfs = "^2025.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
trio = "^0.29.0"

[tool.pytest.ini_options]
pythonpath = ["src"]

[[tool.poetry.source]]
name = "internal"
url = "https://artifactory.hdfcbank.com/artifactory/api/pypi/emvs-python-virtual/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
